requirements.txt
dnspython==2.3.0
motor==3.1.2
pymongo==4.3.3
fastapi==0.96.0
uvicorn==0.22.0
python-decouple==3.8